+++ src/core/platform/opengl-helpers.h
@@ -27,7 +27,7 @@ GLenum FromPixelFormatToDataType(PixelFormat format);
 GLenum FromPixelFormatToShouldNormalizeData(PixelFormat format);
 bool IsPixelFormatInteger(PixelFormat format);
 GLenum FromPrimitiveTopologyToOpenGL(PrimitiveTopology topology);
-GLenum FromIndexSizeToOpenGL(uint32_t index_size);
+GLenum FromIndexSizeToOpenGL(int64_t index_size);
 
 Rndr::String FromOpenGLDataTypeToString(GLenum value);
 Rndr::String FromOpenGLUsageToString(GLenum value);
